feat: refactor search page
